refactor: standardize board coordinate types and improve documentation
- Introduced a new `MutableBoardCoord` type for in-place updates, enhancing clarity in coordinate handling.
- Updated `BoardCoord` to use a readonly tuple for better type safety.
- Refactored imports across multiple files to utilize the new `MutableBoardCoord` type, ensuring consistency.
- Enhanced documentation to clarify the purpose and usage of the new coordinate types, improving user understanding.
