+++ src/board/board-coords.ts
@@ -5,15 +5,14 @@
  * Lookup tables precompute all 64 squares so hot SAN parsing avoids allocations.
  */
 
+import type { MutableBoardCoord } from '#board/mutable-board-coord';
+
 const FILES = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h'] as const;
 
 export type BoardIndex = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7;
 
-/** Mutable board square `[row, col]` (row 0 = rank 8, col 0 = a-file). */
-export type MutableBoardCoord = [row: number, col: number];
-
 /** Read-only board square — public APIs and {@link Action} coord fields. */
-export type BoardCoord = Readonly<MutableBoardCoord>;
+export type BoardCoord = readonly [row: number, col: number];
 
 /** Write into a reusable coord buffer (internal replay / piece lists). */
 export function writeBoardCoord(coord: MutableBoardCoord, row: number, col: number): void {
