+++ src/board/mutable-board-coord.ts
@@ -1,0 +1,2 @@
+/** Mutable `[row, col]` for in-place updates (piece lists, pooled replay buffers). Internal only. */
+export type MutableBoardCoord = [row: number, col: number];
